requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.8.0
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
gunicorn==21.2.0
pytest==8.3.2
pytest-mock==3.14.0

